Refactor page routing using match.params to allow refresh
